% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rrd.R
\name{describe_rrd}
\alias{describe_rrd}
\title{Describes content of a RRD file.}
\usage{
describe_rrd(filename)
}
\arguments{
\item{filename}{File name}
}
\description{
Describes content of a RRD file.
}
\examples{
rrd_cpu_0 <- system.file("extdata/cpu-0.rrd", package = "Rrd")

describe_rrd(rrd_cpu_0)
}
\seealso{
Other rrd functions: \code{\link{getVal}},
  \code{\link{read_rra}}, \code{\link{read_rrd}}
}
