% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVal.R
\name{getVal}
\alias{getVal}
\title{Get value from RRD}
\usage{
getVal(filename, cf, step, timestamp)
}
\arguments{
\item{filename}{File name}

\item{cf}{The consolidation function that is applied to the data you want to fetch. Must be one of \code{c("AVERAGE", "MIN", "MAX", "LAST")}}

\item{step}{step}

\item{timestamp}{timestamp}
}
\value{
data.frame
}
\description{
Returns a \code{data.frame} row associated with a certain \code{timestamp}.
}
\details{
Reads small "chunks" of the RRA (see the \code{rrd.cacheBlock} constant in source
code) and uses a package-wide read-ahead cache implemented as an \link{environment}
object. The \code{timestamp} and \code{step} arguments are also used to calculate the
row index in the cache to ensure fast indexing (as opposed to looking up by
row name, which is known to be much slower in R). The cache is guaranteed
to store a maximum of \code{rrd.cacheSize} rows per RRA.

Use \code{getVal(filename, cf, step, timestamp)} to get values for a single
timestamp. Uses a package-wide read-ahead cache and calculates index in the
cache from the timestamp for faster indexing.

Returns a single \code{data.frame} row. Ensures no more than \code{rrd.cacheSize rows}
per RRA are kept in the cache.
}
\seealso{
Other rrd functions: \code{\link{importRRD}}
}
