% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRRD.R
\name{importRRD}
\alias{importRRD}
\title{Import RRD}
\usage{
importRRD(filename, cf, start, end, step)
}
\arguments{
\item{filename}{File name}

\item{cf}{The consolidation function that is applied to the data you want to fetch. Must be one of \code{c("AVERAGE", "MIN", "MAX", "LAST")}}

\item{start}{start}

\item{end}{end}

\item{step}{step}
}
\value{
Returns a list of data.frames with labels constructed as "consolidation function" + "step" - e.g. "AVERAGE15".
}
\description{
Finds the RRA that best matches the \code{consolidation function} and the \code{step} and imports all values (from all data stores) in that RRA that are between timestamp \code{start} and \code{end}. Note that \code{start} is not included in the result.

Returns a \code{data.frame} object having the \code{timestamp} and the data stores as columns. The data store names are retrieved from the RRD file and set as the corresponding column names. The timestamps are also used as row names.
}
\details{
Use \code{importRRD(filename)} and \code{importRRD(filename, consolidation function, start, end, step)} to import (a part of) an RRA into a \code{data.frame}. The filename, consolidation function and step parameters uniquely identify an RRA array in the RRD file. Start and end define the time-slice to be retrieved. Start is not included in the result. Look at the documentation for \href{https://oss.oetiker.ch/rrdtool/doc/rrdfetch.en.html}{rrdfetch} for more info.

The returned \code{data.frame} has the timestamp and the data stores as separate columns. The names of the data sources are extracted from the RRD file and set as column names. The timestamps are also used as row names.

\code{importRRD(filename)} first reads the metadata in the RRD and adjusts the parameters accordingly in order to expose all RRAs in their entirety.
}
\seealso{
Other rrd functions: \code{\link{getVal}}
}
