% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rrd-package.R
\docType{package}
\name{Rrd-package}
\alias{Rrd-package}
\title{Import data from a RRD file into R}
\description{
The package uses \code{librrd} to import the numerical data in a rrd database
directly into R data structures without using intermediate formats.
}
\details{
Exposes two functions:
\itemize{
\item \code{\link[=importRRD]{importRRD()}}
\item \code{\link[=getVal]{getVal()}}
}
}
\examples{

\dontrun{
importRRD("/var/lib/ganglia/rrds/__SummaryInfo__/mem_free.rrd", 
          "AVERAGE", 1402231215, 1402232085, 15)
rrd = importRRD("/var/lib/ganglia/rrds/__SummaryInfo__/mem_free.rrd")
rrd$AVERAGE15
rrd[[1]]
rrd$AVERAGE15$timestamp
rrd$AVERAGE15$sum
rrd$AVERAGE15[1, 1]
rrd$AVERAGE15[1, "sum"]
rrd[["AVERAGE15"]]
rrd$AVERAGE60
getVal("/var/lib/ganglia/rrds/__SummaryInfo__/mem_free.rrd", 
       "AVERAGE", 15, 1402231245)
rrd.cache[["/var/lib/ganglia/rrds/__SummaryInfo__/mem_free.rrdAVERAGE15"]]
}


}
\references{
\url{http://oss.oetiker.ch/rrdtool/doc/index.en.html}

\url{https://github.com/pldimitrov/Rrd}
}
\author{
Plamen Dimitrov

Maintainer: Plamen Dimitrov \href{mailto:dim.plamen@gmail.com}{dim.plamen@gmail.com}
}
\keyword{librrd}
\keyword{package}
\keyword{rrd}
\keyword{rrdtool}
