% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R, R/read_rrd.R
\name{importRRD}
\alias{importRRD}
\alias{importRRD}
\title{Deprecated functions.}
\usage{
importRRD(filename, cf = NULL, start = NULL, end = NULL, step = NULL)

importRRD(filename, cf = NULL, start = NULL, end = NULL, step = NULL)
}
\arguments{
\item{filename}{File name}

\item{cf}{The consolidation function that is applied to the data you want to fetch. Must be one of \code{c("AVERAGE", "MIN", "MAX", "LAST")}}

\item{start}{start}

\item{end}{end}

\item{step}{step}
}
\description{
Deprecated functions.

Deprecated functions.
}
