% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rrd-package.R
\docType{package}
\name{Rrd-package}
\alias{Rrd-package}
\title{Import data from a RRD database into R}
\description{
The package uses \code{librrd} to import the numerical data in a rrd database
directly into R data structures without using intermediate formats.
}
\details{
Exposes two functions:
\itemize{
\item \code{\link[=read_rrd]{read_rrd()}} to read an entire RRD file
\item \code{\link[=read_rra]{read_rra()}} to extract a single RRA (round robin archive) from an RRD file
\item \code{\link[=getVal]{getVal()}} to extract a single value
}

For more information on rrdtool and the rrd format please refer to the official rrdtool \href{http://oss.oetiker.ch/rrdtool/doc/index.en.html}{documentation} and \href{http://oss.oetiker.ch/rrdtool/tut/index.en.html}{tutorials}.

You can also read a more in-depth description of the package in a \href{http://plamendimitrov.net/blog/2014/08/09/r-package-for-working-with-rrd-files/}{blog post}.
}
\references{
\url{http://oss.oetiker.ch/rrdtool/doc/index.en.html}

\url{https://github.com/pldimitrov/Rrd}
}
\author{
Plamen Dimitrov

Maintainer: Plamen Dimitrov \href{mailto:dim.plamen@gmail.com}{dim.plamen@gmail.com}
}
\keyword{librrd}
\keyword{package}
\keyword{rrd}
\keyword{rrdtool}
